add a textarea for submit order

+++ app_order/views.py
@@ -1,6 +1,7 @@
 from django.views.generic import DetailView
 from .models import Cart, Order, OrderItem
 from django.shortcuts import redirect
+from django.http import HttpResponse
 
 
 class CartDetail(DetailView):
@@ -18,11 +19,14 @@ class CartDetail(DetailView):
 
 
 def submit_order(request):
-    cart_items = request.user.cart.cart_item.all_unique()
-    order = Order.objects.create(cart=request.user.cart)
-    for cart_item in cart_items:
-        OrderItem.objects.create(order=order, shop_product=cart_item.shop_product,
-                                 price=cart_item.shop_product.price,
-                                 count=cart_item.count_same.count())
-    request.user.cart.cart_item.all().delete()
-    return redirect('account:user_profile')
+    if request.method == 'POST':
+        cart_items = request.user.cart.cart_item.all_unique()
+        description = request.POST.get('description')
+        order = Order.objects.create(cart=request.user.cart, description=description)
+        for cart_item in cart_items:
+            OrderItem.objects.create(order=order, shop_product=cart_item.shop_product,
+                                     price=cart_item.shop_product.price,
+                                     count=cart_item.count_same.count())
+        request.user.cart.cart_item.all().delete()
+        return redirect('account:user_profile')
+    return HttpResponse('Bad Request', status=400)
